# Add incremental optimization levels