This is just draft, for discussion. Here i add two new levels of optimization (--level2 and --level1) to existing ones `--release` and `default`.
Reason: I have quite big project which is compiled very slow, and also have big start time (it read data from db, and do some heavy calculations on it). So if i compile it without optimization it starts very slow, if i compile it with --release, compilation take too long. So debugging such project is real pain. By adding new incremental optimization level, i can solve this problem.

Of course this level2, and level1 optimization not even close in terms of performance to `--release` option, because it optimize every module (which is class in crystal), unlike `--release` which optimize united module (using hard inlining). But it fast enough to debug my project.

This is results for my project:
`--release`: initial compile: 61s, incremental compile(change 1 file): 61s, start time: 2s
`--level2`: initial compile: 12.6s, incremental compile(change 1 file): 5.3s, start time: 7s
`--level1`: initial compile: 12s, incremental compile(change 1 file): 5.2s, start time: 7.5s
`default`: initial compile: 7,4s, incremental compile(change 1 file): 5.3s, start time: 23.5s

Build crystal compiler (run time here is recompile crystal by new binary with clean cache and level0, to reduce llvm interference): 
`--release`: initial compile: 7m38,818s, incremental compile(change 1 file): 7m34,734s, run time: 0m36,086s
`--level2`: initial compile: 1m5,084s, incremental compile(change 1 file): 0m22,129s, run time: 0m42,691s
`--level1`: initial compile: 1m1,384s, incremental compile(change 1 file): 0m22,106s, run time: 0m42,980s
`default`: initial compile: 0m27,484s, incremental compile(change 1 file): 0m22,071s, run time: 0m55,572s