@Blacksmoke16 2023-05-12
Related: https://forum.crystal-lang.org/t/faster-release-compile-times-but-slightly-worse-performance/3864