@funny-falcon 2023-05-12
I suggest following options:
- `--no-opt` - matches to current `<default>` mode (and `--level0` in 099c08007ec )
- `<default>` - optimization with O1 and separate/incremental compilation (`--level1`)
- `--opt` - optimization with O2 and separate/incremental compilation (`--level2`)
- `--release` - remains same: O3 and "single_module" compilation

I strongly believe, default mode should be with optimizations enabled, since it is most user uses first. Given it doesn't harm compilation time and provides significant improvement to performance of resulting binary, I don't see why non-optimized mode should remain default.