@kostya 2023-05-13
@jkthorne
> What about more following the optimization levels of compilers like [GCC](https://gcc.gnu.org/onlinedocs/gcc/Optimize-Options.html) and [LLVM](https://clang.llvm.org/docs/CommandGuide/clang.html#code-generation-options)?
> 
> Instead of "--level2" it would be "-O2"?

This is would be quite bad because create confusion, level2 here is not even close to gcc -O2. In gcc -O2 is very good level of optimization, but in crystal it would be much much slower, because use separate module compilation (so no inlining).