@zw963 2023-05-13
I propose use --level1 or --level2 instead of default, because for a project, the first time build time is always can be ignored, because, we only need it to be do once, right?

But, we should keep old default mode for user use it manually.