@kostya 2023-10-31
why not merge it? it not change current compilation (default and --release), only add more options for build customization.