@funny-falcon 2023-11-01
Yeah, quite strange unwilling to improve user experience.