@straight-shoota 2023-11-03
I'm sorry this has been sitting for so long. It's not unwillingness. There's a lot of review work and limited resources. Sometimes PRs fall through the crack. 😢 
Thanks for calling for attention on this. This is definitely one of the contributions that should not be neglected.