@zw963 2025-05-23
> I like idea to use level1 as default.
> 
> minuses:
> 
>     1. 1.5-2 times slower initial compile (done once). For those who use crystal for scripting need to use --level0 option manually.
> 
>     2. less backtrace, need to compile with --debug option, to get same backtrace as default. (adding this option make compilation slower by 15%)
> 
> 
> pluses:
> 
>     1. much faster run time, good for debug big applications, good for big amount of specs.
> 
>     2. similar speed of incremental compile as in default. Most of time we spend in incremental compile.

I have a idea, maybe, we should let `shards` always use -O1 as default. (crystal keep not change)

Because, in most cases, shards are used for a repeatedly compiled project, almost same ` incremental compile` time, but far better runtime performance, it should be a good default.